backends/python/requirements.txt
flask==1.1.2
flask_inputs==0.3.0
solana==0.20.0
redis==4.1.0
base58==2.1.1
pyjwt==2.3.0
flask-cors==3.0.10
python-dotenv==0.19.2